+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add support for lifetimes in relations and the `ref` keyword (#9)
+- Add benchmarks and support for custom hashers (#9)
+
 ## 0.1.3 - 2020-11-21
 
 - Add shorter syntax for defining fact-rules (#6)